Fix Archolos accelerated world clock for scroll quest
Apply SCALETIME to the native world timer and honor the saved hold-time flag. Replay the protected quest checkpoint through native UI saves to cover accelerated and held time.

+++ game/world/objects/npc.cpp
@@ -3022,7 +3022,7 @@ void Npc::tickRoutine() {
       }
     else if(hnpc->start_aistate!=0) {
       auto endTime = owner.time();
-      endTime.addMilis(uint64_t(gtime(4, 0).toInt()));
+      endTime.addMilis(gtime(4, 0).toInt());
       startState(uint32_t(hnpc->start_aistate), "", endTime, false);
       }
     }
